# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/excel/import/templates": {
      "post": {
        "operationId": "ExcelController_importTemplates",
        "parameters": [
          {
            "name": "cycleId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "excel"
        ]
      }
    },
...
    "/api/v1/excel/import/legacy-results": {
      "post": {
        "operationId": "ExcelController_importLegacyResults",
        "summary": "",
        "description": "YoY: 과거결과(평가자정리 시트) 임포트. cycleId 생략 시 2025 사이클 자동탐색.",
        "parameters": [
          {
            "name": "cycleId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "excel"
        ]
      }
    },
...
    "/api/v1/excel/import/kpi/preview": {
      "post": {
        "operationId": "ExcelController_previewKpi",
        "summary": "",
        "description": "개인별 KPI 양식 미리보기(적재 안 함). 파싱 결과 + 가중치합·오류행.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/KpiImportPreviewDto"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "excel"
        ]
      }
    },
    "/api/v1/excel/import/kpi": {
      "post": {
        "operationId": "ExcelController_importKpi",
        "summary": "",
        "description": "개인별 KPI 양식 적재(draft 생성). userId 필수, cycleId 생략 시 활성 사이클.\n멱등: 같은 (userId, cycleId) draft 삭제 후 재생성.",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cycleId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "excel"
        ]
      }
    },
    "/api/v1/excel/import/kpi/commit": {
      "post": {
        "operationId": "ExcelController_commitKpi",
        "summary": "",
        "description": "개인별 KPI — 화면에서 편집한 행 적재(JSON body, multipart 아님).\n미리보기 후 관리자가 정성/정량 토글·누락 보완한 rows 를 그대로 draft 적재.\n멱등: 같은 (userId, cycleId) draft 삭제 후 재생성.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KpiImportCommitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/KpiImportResultDto"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "excel"
        ]
      }
    },
    "/api/v1/excel/import/kpi/submit": {
      "post": {
        "operationId": "ExcelController_submitImportedKpi",
        "summary": "",
        "description": "적재된 draft KPI 제출(2단계 — 적재 후 제출). HR 전용.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KpiImportSubmitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/KpiImportSubmitResultDto"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "excel"
        ]
      }
    },
    "/api/v1/excel/template/{kind}": {
      "get": {
        "operationId": "ExcelController_downloadTemplate",
        "parameters": [
          {
            "name": "kind",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "excel"
        ]
      }
    },
    "/api/v1/excel/export/audit": {
      "get": {
        "operationId": "ExcelController_exportAudit",
        "parameters": [
          {
            "name": "actorId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entity",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "excel"
        ]
      }
    },
    "/api/v1/excel/export/results": {
      "get": {
        "operationId": "ExcelController_exportResults",
        "parameters": [
          {
            "name": "cycleId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "excel"
        ]
      }
    },
    "/api/v1/excel/export/distribution": {
      "get": {
        "operationId": "ExcelController_exportDistribution",
        "parameters": [
          {
            "name": "cycleId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "excel"
        ]
      }
    },
    "/api/v1/excel/export/compensation": {
      "get": {
        "operationId": "ExcelController_exportCompensation",
        "parameters": [
          {
            "name": "cycleId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "excel"
        ]
      }
    },
...
    "/api/v1/monthly-performance": {
      "get": {
        "operationId": "MonthlyPerformanceController_list",
        "parameters": [
          {
            "name": "cycleId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MonthlyPerformanceDto"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "number"
                        },
                        "pageSize": {
                          "type": "number"
                        },
                        "total": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "monthly-performance"
        ]
      },
      "post": {
        "operationId": "MonthlyPerformanceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMonthlyPerformanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/MonthlyPerformanceDto"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "monthly-performance"
        ]
      }
    },
    "/api/v1/monthly-performance/summary": {
      "get": {
        "operationId": "MonthlyPerformanceController_summary",
        "parameters": [
          {
            "name": "cycleId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/MonthlyPerformanceSummaryDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "monthly-performance"
        ]
      }
    },
    "/api/v1/monthly-performance/{id}": {
      "patch": {
        "operationId": "MonthlyPerformanceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMonthlyPerformanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/MonthlyPerformanceDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "monthly-performance"
        ]
      }
    },
...
  },
  "info": {
    "title": "에너지엑스 인사평가 API",
    "description": "모듈러 모놀리식 백엔드 — 응답 봉투 {data}/{data,meta}/{error}",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "KpiImportRowDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "revenue",
              "construction",
              "orders",
              "collaboration",
              "development"
            ]
          },
          "group": {
            "type": "string",
            "enum": [
              "performance_core",
              "collaboration_growth"
            ]
          },
          "csf": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "targetText": {
            "type": "string",
            "nullable": true
          },
          "measureMethod": {
            "type": "string",
            "nullable": true
          },
          "weight": {
            "type": "number",
            "nullable": true
          },
          "isQualitative": {
            "type": "boolean"
          },
          "gradingCriteria": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "type": "string"
            },
            "description": "정성 등급기준 {S,A,B,C,D} 텍스트"
          },
          "valid": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "category",
          "group",
          "csf",
          "title",
          "targetText",
          "measureMethod",
          "weight",
          "isQualitative",
          "gradingCriteria",
          "valid",
          "message"
        ]
      },
      "KpiImportErrorDto": {
        "type": "object",
        "properties": {
          "row": {
            "type": "number"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "row",
          "message"
        ]
      },
      "KpiImportPreviewDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KpiImportRowDto"
            }
          },
          "validCount": {
            "type": "number"
          },
          "errorCount": {
            "type": "number"
          },
          "weightSum": {
            "type": "number"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KpiImportErrorDto"
            }
          }
        },
        "required": [
          "fileName",
          "rows",
          "validCount",
          "errorCount",
          "weightSum",
          "errors"
        ]
      },
      "KpiImportResultDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          },
          "cycleId": {
            "type": "string"
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "imported": {
            "type": "number"
          },
          "deletedDrafts": {
            "type": "number"
          },
          "weightSum": {
            "type": "number"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KpiImportErrorDto"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ok",
          "userId",
          "cycleId",
          "fileName",
          "imported",
          "deletedDrafts",
          "weightSum",
          "errors",
          "warnings"
        ]
      },
      "KpiImportGradingCriteriaDto": {
        "type": "object",
        "properties": {
          "S": {
            "type": "string",
            "nullable": true
          },
          "A": {
            "type": "string",
            "nullable": true
          },
          "B": {
            "type": "string",
            "nullable": true
          },
          "C": {
            "type": "string",
            "nullable": true
          },
          "D": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "KpiImportCommitRowDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "object"
          },
          "group": {
            "type": "object"
          },
          "csf": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "description": "성과관리지표(KPI). 빈 행은 스킵 + warning."
          },
          "targetText": {
            "type": "string",
            "nullable": true
          },
          "measureMethod": {
            "type": "string",
            "nullable": true
          },
          "weight": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "isQualitative": {
            "type": "boolean",
            "description": "관리자 토글(정성/정량)."
          },
          "gradingCriteria": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/KpiImportGradingCriteriaDto"
              }
            ]
          }
        },
        "required": [
          "category",
          "group",
          "title",
          "weight",
          "isQualitative"
        ]
      },
      "KpiImportCommitDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "대상 사용자(필수)."
          },
          "cycleId": {
            "type": "string",
            "description": "생략 시 활성 사이클(없으면 400)."
          },
          "fileName": {
            "type": "string"
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KpiImportCommitRowDto"
            }
          }
        },
        "required": [
          "userId",
          "rows"
        ]
      },
      "KpiImportSubmitResultDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          },
          "cycleId": {
            "type": "string"
          },
          "submitted": {
            "type": "number"
          },
          "weightSum": {
            "type": "number"
          }
        },
        "required": [
          "ok",
          "userId",
          "cycleId",
          "submitted",
          "weightSum"
        ]
      },
      "KpiImportSubmitDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "대상 사용자(필수)."
          },
          "cycleId": {
            "type": "string",
            "description": "생략 시 활성 사이클."
          }
        },
        "required": [
          "userId"
        ]
      },
...
      "MonthlyPerformanceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "cycleId": {
            "type": "string"
          },
          "departmentId": {
            "type": "string"
          },
          "year": {
            "type": "number"
          },
          "month": {
            "type": "number",
            "description": "1~12."
          },
          "category": {
            "type": "string",
            "enum": [
              "revenue",
              "construction",
              "orders",
              "collaboration",
              "development"
            ]
          },
          "targetAmount": {
            "type": "number"
          },
          "actualAmount": {
            "type": "number"
          },
          "achievementRate": {
            "type": "number",
            "description": "단월 달성률(actual/target×100, 목표 0이면 0)."
          },
          "enteredById": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "cycleId",
          "departmentId",
          "year",
          "month",
          "category",
          "targetAmount",
          "actualAmount",
          "achievementRate",
          "enteredById",
          "createdAt",
          "updatedAt"
        ]
      },
      "MonthlyPerformanceSummaryCategoryDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "revenue",
              "construction",
              "orders",
              "collaboration",
              "development"
            ]
          },
          "targetAmount": {
            "type": "number"
          },
          "actualAmount": {
            "type": "number"
          },
          "achievementRate": {
            "type": "number"
          },
          "currentGrade": {
            "type": "string",
            "nullable": true,
            "description": "측정방식별 등급(S~D). 실적 미입력 시 null."
          }
        },
        "required": [
          "category",
          "targetAmount",
          "actualAmount",
          "achievementRate",
          "currentGrade"
        ]
      },
      "MonthlyTrendPointDto": {
        "type": "object",
        "properties": {
          "month": {
            "type": "number",
            "description": "1~12."
          },
          "achievementRate": {
            "type": "number"
          },
          "grade": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "month",
          "achievementRate",
          "grade"
        ]
      },
      "MonthlyPerformanceSummaryDto": {
        "type": "object",
        "properties": {
          "cycleId": {
            "type": "string"
          },
          "departmentId": {
            "type": "string"
          },
          "departmentName": {
            "type": "string",
            "nullable": true
          },
          "targetAmount": {
            "type": "number",
            "description": "누적 목표 합."
          },
          "actualAmount": {
            "type": "number",
            "description": "누적 실적 합."
          },
          "achievementRate": {
            "type": "number"
          },
          "currentGrade": {
            "type": "string",
            "nullable": true,
            "description": "종합 등급(S~D). 실적 미입력 시 null."
          },
          "byCategory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonthlyPerformanceSummaryCategoryDto"
            }
          },
          "monthlyTrend": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonthlyTrendPointDto"
            }
          }
        },
        "required": [
          "cycleId",
          "departmentId",
          "departmentName",
          "targetAmount",
          "actualAmount",
          "achievementRate",
          "currentGrade",
          "byCategory",
          "monthlyTrend"
        ]
      },
      "CreateMonthlyPerformanceDto": {
        "type": "object",
        "properties": {
          "cycleId": {
            "type": "string"
          },
          "departmentId": {
            "type": "string"
          },
          "year": {
            "type": "number",
            "minimum": 2000,
            "maximum": 2100
          },
          "month": {
            "type": "number",
            "minimum": 1,
            "maximum": 12
          },
          "category": {
            "type": "object"
          },
          "targetAmount": {
            "type": "number"
          },
          "actualAmount": {
            "type": "number"
          }
        },
        "required": [
          "cycleId",
          "departmentId",
          "year",
          "month",
          "category",
          "targetAmount",
          "actualAmount"
        ]
      },
      "UpdateMonthlyPerformanceDto": {
        "type": "object",
        "properties": {
          "targetAmount": {
            "type": "number"
          },
          "actualAmount": {
            "type": "number"
          }
        }
      },
...
    }
  }
}