# packages/contracts/src/generated/model/monthlyTrendPointDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * 에너지엑스 인사평가 API
 * 모듈러 모놀리식 백엔드 — 응답 봉투 {data}/{data,meta}/{error}
 * OpenAPI spec version: 1.0.0
 */

export interface MonthlyTrendPointDto {
  /** 1~12. */
  month: number;
  achievementRate: number;
  /** @nullable */
  grade: string | null;
}